% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUBClustR_functions.R
\name{calculate_WAD}
\alias{calculate_WAD}
\title{Calculate Weighted Average Dissimilarity (WAD) and Detailed Dissimilarities}
\usage{
calculate_WAD(model_obj, data, m)
}
\arguments{
\item{model_obj}{The model object returned by \code{fit_MLCCUB}.}

\item{data}{The original dataset containing the observed ordinal responses.}

\item{m}{A numeric vector specifying the maximum number of categories for each item.}
}
\value{
A list containing:
\itemize{
\item \code{WAD}: Global Weighted Average Dissimilarity index.
\item \code{Cluster_Dissimilarity}: Unweighted average dissimilarity for each cluster.
\item \code{Item_Dissimilarity}: Weighted average dissimilarity for each item.
\item \code{Matrix_Dissimilarity}: A K x J matrix of normalized dissimilarities.
}
}
\description{
Computes the WAD index and detailed dissimilarities to assess the
goodness of fit of the MLC-CUB model. Returns values in \verb{[0, 1]}, where 0
indicates a perfect fit between observed data and theoretical probabilities.
}
\examples{
\dontrun{
set.seed(123)
data_mock <- matrix(sample(1:5, 100 * 4, replace = TRUE), ncol = 4)
m_vec <- rep(5, 4)
model <- fit_MLCCUB(data_mock, m = m_vec, K = 2, EM_iter = 1, max_iter = 20)

# Calculate the Weighted Average Dissimilarity (WAD) indices
wad_metrics <- calculate_WAD(model_obj = model, data = data_mock, m = m_vec)

# Print the global WAD index (values closer to 0 indicate a better fit)
print(wad_metrics$WAD)
}
}
