% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUBClustR_functions.R
\name{check_identifiability}
\alias{check_identifiability}
\title{Model Identifiability Check via Bootstrap}
\usage{
check_identifiability(
  data,
  m,
  K,
  n_boot = 100,
  n_cores = 1,
  plot_hist = TRUE,
  seed = 250535,
  cpp_file = .cpp_path,
  ...
)
}
\arguments{
\item{data}{The original dataset (matrix or data.frame).}

\item{m}{A numeric vector specifying the maximum number of categories for each item.}

\item{K}{The number of clusters to test.}

\item{n_boot}{Number of bootstrap samples to extract (default: 100).}

\item{n_cores}{Number of CPU cores for parallel computation (default: 1 for sequential).}

\item{plot_hist}{Logical. If TRUE, generates and returns the ARI distribution histogram.}

\item{seed}{Random seed for reproducibility.}

\item{cpp_file}{Path to the C++ engine file to be loaded by workers (default: "engine.cpp").}

\item{...}{Additional arguments passed to \code{fit_MLCCUB} (e.g., \code{EM_iter}, \code{tol}).}
}
\value{
A list containing the pairwise ARI matrix (\code{ARI_matrix}), the flattened ARI values (\code{ARI_values}), the ggplot object (\code{plot}), and the raw cluster assignments (\code{raw_clusters}).
}
\description{
Performs a non-parametric bootstrap to assess the stability and identifiability of the latent clusters using the Adjusted Rand Index (ARI).
}
\examples{
\dontrun{
set.seed(123)
data_mock <- matrix(sample(1:5, 100 * 4, replace = TRUE), ncol = 4)

# Fast bootstrap with only 5 iterations and 1 core (for testing purposes)
ident_res <- check_identifiability(data = data_mock, m = rep(5, 4), K = 2, 
                                   n_boot = 5, n_cores = 1, EM_iter = 1)

# Display the histogram of the Adjusted Rand Index (ARI) distribution
print(ident_res$plot)
}
}
